pyproject.toml
[project]
name = "grpc-2"
version = "0.1.0"
description = ""
authors = [
    {name = "olga.baklanova",email = "o.lgaromanovna@yandex.ru"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "grpcio-tools (>=1.73.1,<2.0.0)",
    "protobuf (>=6.31.1,<7.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
